+++ CHANGELOG.md
@@ -132,6 +132,7 @@
 - EXB-1046: Don't delete a channel used in an "Export to Shared Catalogs" export profile
 - PIM-10087: Fix storage errors HTTP code to return 500 instead of 422
 - PIM-10090: Fix missing cache clearing during family variant changes computing
+- PIM-10048: fix memory leak in search product models by family variant query
 
 ## New features